Move reducers test to dedicated file
